+++ packages/orchestration/src/facade.js
@@ -53,42 +53,7 @@ export const makeLocalChainFacade = (
         storageNode: null,
       });
 
-      // FIXME turn this into an Exo LocalChainOrchestrationAccount or make that a facet of makeLocalChainAccountKit
-      return {
-        async deposit(payment) {
-          console.log('deposit got', payment);
-          await E(account).deposit(payment);
-        },
-        getAddress() {
-          return account.getAddress();
-        },
-        async getBalance(denomArg) {
-          // FIXME look up real values
-          // UNTIL https://github.com/Agoric/agoric-sdk/issues/9211
-          const [brand, denom] =
-            typeof denomArg === 'string'
-              ? [/** @type {any} */ (null), denomArg]
-              : [denomArg, 'FIXME'];
-
-          const natAmount = await E(lca).getBalance(brand);
-          return harden({ denom, value: natAmount.value });
-        },
-        getBalances() {
-          throw new Error('not yet implemented');
-        },
-        async send(toAccount, amount) {
-          // FIXME implement
-          console.log('send got', toAccount, amount);
-        },
-        async transfer(amount, destination, opts) {
-          console.log('transfer got', amount, destination, opts);
-          return account.transfer(amount, destination, opts);
-        },
-        transferSteps(amount, msg) {
-          console.log('transferSteps got', amount, msg);
-          return Promise.resolve();
-        },
-      };
+      return account;
     },
   });
 };
